`dowords` should only pass valid nametokens
`dowords` checked for the last nametoken *after* calling the lambda.
This meant that the lambda would get called on the sentinel byte at the
end of the wordlist.

Moved the last-nametoken check to before the lambda, saving the name
length on the stack so that we don't have to read it again at the end of
the loop.