Forth word index in latex documentation (#459)
* Add index to manual.

Index entries in next commit.

* Some index entries (not thoroughly proofread).

Renders a pretty index with page references in the PDF, but has
some clear errors, like 2@ showing up as 2 in its own section.

# Conflicts:
#	docs/words.tex

* Correct quoting of " and @ for index

Quote characters by prefixing "

* More " quoting of index entries

* Use printf instead of echo for consistency

Turns out POSIX just left backslash as implementation defined for echo.

Many versions of echo have -e/-E switches to decide if escape sequences
are replaced or not, but not all. In the case of dash, which is /bin/sh
on some Debian systems, escape sequences can't be turned off.

printf, on the other hand, does have specified behaviour in that it
will expand escape sequences in the format string, but not the
%s arguments.

* Some index tag tuning.

For epub, Foliate can show a preview of the line with the index tag.
This made \index point poorly when placed before \item.
However, epub still has issues, like actually jumping to the section
instead of the particular paragraph, and showing some strange numbers.
Worst, the index table breaks in long "pages". Span the headings?

# Conflicts:
#	docs/words.tex

* make4ht build script

* Build html and epub using tex4ht.

Quite hacky index formatting (only Foliate shows the columns,
yet the links only go to the chapter despite previewing the
correct line).

# Conflicts:
#	docs/Makefile
#	docs/manual.tex

* Clarify purpose of manual.ist index style